Add catalog read path: search, show detail, episodes
Implement the catalog API from spec/03-api against TMDB:
- GET /api/search: proxy TMDB search, annotate shows we've imported.
- GET /api/shows/{id}: show + seasons; {id} is our uuid or tmdb:<n> to
  import-on-demand (series + seasons + episodes), per spec/04.
- GET /api/shows/{id}/episodes: episodes with aggregate filler scores derived
  via the scoring module.
- Cache-Control on catalog responses (long for catalog, short for vote-derived
  scores) per ADR-0005; JSON error shape + proper 400/404/502/503 mapping.

Uses sqlx compile-time-checked queries with a committed .sqlx offline cache, so
builds/CI need no database. Verified online + offline (15 tests) and via a live
smoke test against Postgres (show detail, episode scores, error cases).

Reviewed by a separate agent before commit per project workflow; fixed its
blocker (made import transactional with all TMDB fetches up front, so a mid-
import failure leaves no partial state), stopped health/db leaking the DB error
string, and documented the episode-renumbering re-import gap for the refresh
milestone.

+++ api/migrations/0001_init.sql
@@ -7,7 +7,7 @@ CREATE TYPE vote_value AS ENUM ('FILLER', 'CANON');
 
 CREATE TABLE show (
     id             UUID PRIMARY KEY DEFAULT gen_random_uuid(),
-    tmdb_id        INTEGER NOT NULL UNIQUE,
+    tmdb_id        BIGINT NOT NULL UNIQUE,
     name           TEXT NOT NULL,
     first_air_year INTEGER,
     poster_path    TEXT,
@@ -28,7 +28,7 @@ CREATE TABLE episode (
     id             UUID PRIMARY KEY DEFAULT gen_random_uuid(),
     show_id        UUID NOT NULL REFERENCES show (id) ON DELETE CASCADE,
     season_id      UUID NOT NULL REFERENCES season (id) ON DELETE CASCADE,
-    tmdb_id        INTEGER NOT NULL UNIQUE,
+    tmdb_id        BIGINT NOT NULL UNIQUE,
     season_number  INTEGER NOT NULL,
     episode_number INTEGER NOT NULL,
     name           TEXT,
